trade-trigger/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
cachetools==5.2.0
cffi==1.15.1
charset-normalizer==2.1.1
cloud-sql-python-connector==0.9.3
cryptography==38.0.3
DateTime==4.7
frozenlist==1.3.3
google==3.0.0
google-auth==2.14.1
greenlet==2.0.1
idna==3.4
multidict==6.0.2
pg8000==1.29.3
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
rsa==4.9
scramp==1.4.4
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.43
urllib3==1.26.12
yarl==1.8.1
zope.interface==5.5.1
